Update routes, Move LandingPage and NotFoundPage to pages folder
